Renamed example folder to client folder
