--- example/about.html
@@ -1,116 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <style type="text/css">
-        body {
-            font-family: 'Helvetica Neue', Helvetica, Arial, 'Lucida Grande', sans-serif;
-            background-image:url(https://images.pexels.com/photos/802958/pexels-photo-802958.jpeg?cs=srgb&dl=earphones-earpods-justifyyourlove-802958.jpg&fm=jpg);
-            /*height: 100%;*/
-            background-position: center;
-            background-repeat: no-repeat;
-            background-size: cover;
-        }
-
-        .user {
-            font-weight: 200;
-            font-size: 20px;
-        }
-
-        ul {
-            list-style-type: none;
-            margin: 0;
-            padding: 0;
-            overflow: hidden;
-            background-color: rgb(80, 135, 160);
-            position: fixed;
-            top: 0;
-            width: 100%;
-        }
-
-        li {
-            float: left;
-        }
-
-        li a {
-            display: block;
-            color: white;
-            text-align: center;
-            padding: 14px 16px;
-            text-decoration: none;
-        }
-
-        li a:hover {
-            background-color: rgba(0, 0, 0, 0.40);
-        }
-
-        .active {
-            background-color: rgba(0, 0, 0, 0.40);
-        }
-
-        div.gallery {
-            margin: 5px;
-            border: 1px solid #ccc;
-            background-color: rgba(105, 105, 105, 0.35);
-            float: left;
-            width: 180px;
-        }
-
-        div.gallery img {
-            width: 100%;
-            height: auto;
-        }
-
-        div.desc {
-            padding: 15px;
-            text-align: center;
-        }
-    </style>
-</head>
-<body>
-</body>
-    <br>
-    <br>
-    <ul>
-        <li><a href="client.html" style="font-weight: bold; text-align:right;">TrackHub</a></li>
-        <li><a class="active" href="about.html">About</a></li>
-        <li><a href="contact.html">Contact</a></li>
-        <li style="float:right"><a href="register.html" >Register</a></li>
-        <li style="float:right"><a href="login.html">Login</a></li>
-    </ul>
-
-<div class="subheader">
-    <br>
-    <div class="subhead-text" style="font-size: 28px; font-family:tahoma; color: antiquewhite; text-align: left; text-decoration: underline;">About Us</div><br><span class="user"></span></div>
-</div>
-
-<h4 style=" color: darkslategray; font-family:tahoma; font-size: 22px;">The Project:</h4>
-<p style="color: white; font-family:tahoma;">TrackHub is an interactive music player that allows users to create an account, stream music, rate songs, chat with other members, and allows users to see a history of their songs. One of our goals when creating TrackHub was to make it cross platform, so we implemented it as a website available to run in any modern web browser. The main features that distinguish TrackHub from other music players are its rating/review, chat, and music video integration. .</p></br>
-<h4 style=" color: darkslategray; font-family:tahoma; font-size: 22px;">The Team:</h4>
-
-<div class="gallery">
-    <img src="/assets/Blank.jpg" alt="Edgar" style="width:100%"/>
-    <div class="desc"; style="color:white; font-weight: bold;">Edgar Oregel</div>
-</div>
-
-<div class="gallery">
-    <img src="/assets/ClementG.JPG" alt="Clement" style="width:100%"/>
-    <div class="desc" style="color:white; font-weight: bold;">Clement Garcia</div>
-</div>
-
-<div class="gallery">
-    <img src="/assets/DrewO.png" alt="Drew" style="width:100%"/>
-    <div class="desc" style="color:white; font-weight: bold;">Drew Overgaard</div>
-</div>
-
-<div class="gallery">
-    <img src="/assets/KyleH.JPG" alt="Kyle" style="width:100%"/>
-    <div class="desc" style="color:white; font-weight: bold;">Kyle Hartman</div>
-</div>
-
-<div class="gallery">
-    <img src="/assets/Blank.jpg" alt="Ryan" style="width:100%"/>
-    <div class="desc" style="color:white; font-weight: bold;">Ryan Espinosa</div>
-</div>
-
-</body>
-</html>
