--- example/login.html
@@ -1,163 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <title>Login V5</title>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <!--===============================================================================================-->
-    <link rel="icon" type="image/png" href="images/icons/favicon.ico"/>
-    <!--===============================================================================================-->
-    <link rel="stylesheet" type="text/css" href="vendor/bootstrap/css/bootstrap.min.css">
-    <!--===============================================================================================-->
-    <link rel="stylesheet" type="text/css" href="fonts/font-awesome-4.7.0/css/font-awesome.min.css">
-    <!--===============================================================================================-->
-    <link rel="stylesheet" type="text/css" href="fonts/Linearicons-Free-v1.0.0/icon-font.min.css">
-    <!--===============================================================================================-->
-    <link rel="stylesheet" type="text/css" href="vendor/animate/animate.css">
-    <!--===============================================================================================-->
-    <link rel="stylesheet" type="text/css" href="vendor/css-hamburgers/hamburgers.min.css">
-    <!--===============================================================================================-->
-    <link rel="stylesheet" type="text/css" href="vendor/animsition/css/animsition.min.css">
-    <!--===============================================================================================-->
-    <link rel="stylesheet" type="text/css" href="vendor/select2/select2.min.css">
-    <!--===============================================================================================-->
-    <link rel="stylesheet" type="text/css" href="vendor/daterangepicker/daterangepicker.css">
-    <!--===============================================================================================-->
-    <link rel="stylesheet" type="text/css" href="css/util.css">
-    <link rel="stylesheet" type="text/css" href="css/main.css">
-    <!--===============================================================================================-->
-    <style type="text/css">
-        body {
-            font-family: 'Helvetica Neue', Helvetica, Arial, 'Lucida Grande', sans-serif;
-            background-image:url(https://images.pexels.com/photos/802958/pexels-photo-802958.jpeg?cs=srgb&dl=earphones-earpods-justifyyourlove-802958.jpg&fm=jpg);
-            /*height: 100%;*/
-            background-position: center;
-            background-repeat: no-repeat;
-            background-size: cover;
-        }
-
-        .user {
-            font-weight: 200;
-            font-size: 20px;
-        }
-
-        ul {
-            list-style-type: none;
-            margin: 0;
-            padding: 0;
-            overflow: hidden;
-            background-color: darkslategray;
-        }
-
-        li {
-            float: left;
-        }
-
-        li a {
-            display: block;
-            color: white;
-            text-align: center;
-            padding: 14px 16px;
-            text-decoration: none;
-        }
-
-        li a:hover {
-            background-color: rgba(105, 105, 105, 0.75);
-        }
-    </style>
-</head>
-<body>
-
-<ul>
-    <!--<li><img src="/assets/headphones.svg" alt="headphones" style="width:10%"/></li>-->
-    <li><a href="client.html" style="font-size: 20px; font-weight: bold; text-align:right;">TrackHub</a></li>
-    <li><a href="about.html">About </a></li>
-    <li><a href="contact.html">Contact</a></li>
-    <li><a class="active" href="login.html">Login</a></li>
-    <li><a href="register.html" >Register</a></li>
-</ul>
-
-<div class="limiter">
-    <div class="container-login100" style="background-image: url('https://images.pexels.com/photos/802958/pexels-photo-802958.jpeg?cs=srgb&dl=earphones-earpods-justifyyourlove-802958.jpg&fm=jpg');">
-        <div class="wrap-login100 p-l-110 p-r-110 p-t-62 p-b-33">
-            <form class="login100-form validate-form flex-sb flex-w">
-					<span class="login100-form-title p-b-53">
-						Sign In With
-					</span>
-
-                <a href="#" class="btn-face m-b-20">
-                    <i class="fa fa-facebook-official"></i>
-                    Facebook
-                </a>
-
-                <a href="#" class="btn-google m-b-20">
-                    <img src="images/icons/icon-google.png" alt="GOOGLE">
-                    Google
-                </a>
-
-                <div class="p-t-31 p-b-9">
-						<span class="txt1">
-							Username
-						</span>
-                </div>
-                <div class="wrap-input100 validate-input" data-validate = "Username is required">
-                    <input class="input100" type="text" name="username" >
-                    <span class="focus-input100"></span>
-                </div>
-
-                <div class="p-t-13 p-b-9">
-						<span class="txt1">
-							Password
-						</span>
-
-                   <!--<a href="#" class="txt2 bo1 m-l-5">
-                        Forgot?
-                    </a>-->
-                </div>
-                <div class="wrap-input100 validate-input" data-validate = "Password is required">
-                    <input class="input100" type="password" name="pass" >
-                    <span class="focus-input100"></span>
-                </div>
-
-                <div class="container-login100-form-btn m-t-17">
-                    <button class="login100-form-btn">
-                        Sign In
-                    </button>
-                </div>
-
-                <div class="w-full text-center p-t-55">
-						<span class="txt2">
-							Not a member?
-						</span>
-
-                    <a href="register.html" class="txt2 bo1">
-                        Sign up now
-                    </a>
-                </div>
-            </form>
-        </div>
-    </div>
-</div>
-
-
-<div id="dropDownSelect1"></div>
-
-<!--===============================================================================================-->
-<script src="vendor/jquery/jquery-3.2.1.min.js"></script>
-<!--===============================================================================================-->
-<script src="vendor/animsition/js/animsition.min.js"></script>
-<!--===============================================================================================-->
-<script src="vendor/bootstrap/js/popper.js"></script>
-<script src="vendor/bootstrap/js/bootstrap.min.js"></script>
-<!--===============================================================================================-->
-<script src="vendor/select2/select2.min.js"></script>
-<!--===============================================================================================-->
-<script src="vendor/daterangepicker/moment.min.js"></script>
-<script src="vendor/daterangepicker/daterangepicker.js"></script>
-<!--===============================================================================================-->
-<script src="vendor/countdowntime/countdowntime.js"></script>
-<!--===============================================================================================-->
-<script src="js/main.js"></script>
-
-</body>
-</html>
